expense/requirements.txt
Django==5.0.4
djangorestframework
gunicorn
